:recycle: Push PlaceEdit object encoding to PlacEdit model.
